+++ packages/ionic-react/src/schematics/application/schematic.spec.ts
@@ -68,12 +68,12 @@ describe('application', () => {
     ).toBeTruthy();
     expect(
       tree.exists(
-        `${projectRoot}/src/app/pages/${homeFileName}.${componentExtension}`
+        `${projectRoot}/src/pages/${homeFileName}.${componentExtension}`
       )
     ).toBeTruthy();
     expect(
       tree.exists(
-        `${projectRoot}/src/app/components/${exploreContainerFileName}.${componentExtension}`
+        `${projectRoot}/src/components/${exploreContainerFileName}.${componentExtension}`
       )
     ).toBeTruthy();
 
@@ -83,13 +83,11 @@ describe('application', () => {
     ) {
       expect(
         tree.exists(
-          `${projectRoot}/src/app/components/${exploreContainerFileName}.${options.style}`
+          `${projectRoot}/src/components/${exploreContainerFileName}.${options.style}`
         )
       ).toBeTruthy();
       expect(
-        tree.exists(
-          `${projectRoot}/src/app/pages/${homeFileName}.${options.style}`
-        )
+        tree.exists(`${projectRoot}/src/pages/${homeFileName}.${options.style}`)
       ).toBeTruthy();
       expect(
         tree.exists(`${projectRoot}/src/app/theme/variables.${options.style}`)
@@ -97,13 +95,11 @@ describe('application', () => {
     } else {
       expect(
         tree.exists(
-          `${projectRoot}/src/app/components/${exploreContainerFileName}.${options.style}`
+          `${projectRoot}/src/components/${exploreContainerFileName}.${options.style}`
         )
       ).toBeFalsy();
       expect(
-        tree.exists(
-          `${projectRoot}/src/app/pages/${homeFileName}.${options.style}`
-        )
+        tree.exists(`${projectRoot}/src/pages/${homeFileName}.${options.style}`)
       ).toBeFalsy();
       expect(
         tree.exists(`${projectRoot}/src/app/theme/variables.${options.style}`)